Shift handler added

+++ src/main/java/com/bloodbank/bloodbank/controllers/bank/ManagementController.java
@@ -14,7 +14,9 @@ import lombok.Setter;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
+import org.springframework.validation.BeanPropertyBindingResult;
 import org.springframework.validation.BindingResult;
+import org.springframework.validation.FieldError;
 import org.springframework.web.bind.annotation.*;
 
 import java.time.LocalDate;
@@ -36,8 +38,8 @@ public class ManagementController {
 
     @GetMapping("/manage")
     public String manage(Model model) {
-        setModelAttributes(model, new Date(), employeeService.getAll(), shiftService.getAll());
-        return "admin/management";
+        setModelAttributes(model, employeeService.getAll(), shiftService.getAll());
+        return "admin/edittest";
     }
 
     @PostMapping("/manage/em/search")
@@ -53,7 +55,7 @@ public class ManagementController {
             String[] nameParts = s.split(" ");
             employeeList.addAll(employeeService.findByName(nameParts[0], nameParts[1], nameParts[2]));
         }
-        setModelAttributes(model, new Date(), employeeList, shiftService.getAll());
+        setModelAttributes(model, search, new Date(), employeeList, shiftService.getAll());
         return "admin/management";
     }
 
@@ -72,6 +74,16 @@ public class ManagementController {
                 result.rejectValue("difference", "error.difference", "Разница между датами должна быть не менее 16 лет.");
             }
         }
+        if (!employee.getPhone().isBlank() && action.equals("edit")) {
+            List<FieldError> errorsToKeep = result.getFieldErrors().stream()
+                    .filter(fer -> !fer.getField().equals("phone")).toList();
+
+            result = new BeanPropertyBindingResult(employee, "employee");
+
+            for (FieldError fieldError : errorsToKeep) {
+                result.addError(fieldError);
+            }
+        }
         if (result.hasErrors()) {
             return "admin/editEM";
         }
@@ -81,7 +93,6 @@ public class ManagementController {
         if (action.equals("edit")) {
             employeeService.update(employee);
         }
-        setModelAttributes(model, new Date(), employeeService.getAll(), shiftService.getAll());
         return "redirect:/manage";
     }
 
@@ -94,6 +105,8 @@ public class ManagementController {
 
     @RequestMapping(value = "/manage/em/delete", method = RequestMethod.GET)
     public String emDelete(@RequestParam Long id, Model model) {
+        if (shiftService.isEmployeeEngaged(id))
+            return "redirect:/manage?error";
         employeeService.delete(id);
         return "redirect:/manage";
     }
@@ -104,27 +117,40 @@ public class ManagementController {
         List<Shift> shifts;
         if (date.from == null && date.to == null) {
             shifts = shiftService.getAll();
-            setModelAttributes(model, date, employeeService.getAll(), shifts);
+            setModelAttributes(model, new Search(), date, employeeService.getAll(), shifts);
             return "admin/management";
         }
         date.setFrom(date.from == null ? LocalDate.now() : date.from);
-        date.setTo(date.to == null ? LocalDate.now() : date.from);
+        date.setTo(date.to == null ? LocalDate.now() : date.to);
         shifts = shiftService.findAllShiftsWithDateRange(date.from, date.to);
-        setModelAttributes(model, date, employeeService.getAll(), shifts);
+        setModelAttributes(model, new Search(), date, employeeService.getAll(), shifts);
         return "admin/management";
     }
 
     @GetMapping("/manage/sh/add")
     public String shAdd(Model model) {
+        action = "add";
         model.addAttribute("shift", new ShiftDTO());
         model.addAttribute("employees", employeeService.findByName().stream()
                 .map(em -> em[0] + " " + em[1] + " " + em[2] + " " + em[3]).toList());
+        model.addAttribute("searchEm", new Search());
         return "admin/editSH";
     }
 
     @PostMapping("/manage/sh/add")
-    public String shAdd(@Valid @ModelAttribute ShiftDTO shiftDTO, BindingResult bindingResult, Model model) {
-
+    public String shAdd(@Valid @ModelAttribute("shift") ShiftDTO shiftDTO, BindingResult bindingResult, Model model) {
+        if (shiftDTO.getDate() != null && shiftService.isShiftThisDay(shiftDTO.getDate()) && action.equals("add"))
+            bindingResult.rejectValue("date", "error.date", "На эту дату уже назначена смена.");
+        if (bindingResult.hasErrors()) {
+            model.addAttribute("shift", shiftDTO);
+            model.addAttribute("employees", employeeService.findByName().stream()
+                    .map(em -> em[0] + " " + em[1] + " " + em[2] + " " + em[3]).toList());
+            return "admin/editSH";
+        }
+        if (action.equals("add"))
+            shiftService.add(shiftDTO);
+        if (action.equals("edit"))
+            shiftService.update(shiftDTO);
         return "redirect:/manage";
     }
 
@@ -134,13 +160,36 @@ public class ManagementController {
         return "redirect:/manage";
     }
 
+    @RequestMapping("/manage/sh/edit")
+    public String shEdit(@RequestParam Long id, Model model) {
+        action = "edit";
+        model.addAttribute("shift", shiftService.mapToDTO(shiftService.findById(id)));
+        model.addAttribute("employees", employeeService.findByName().stream()
+                .map(em -> em[0] + " " + em[1] + " " + em[2] + " " + em[3]).toList());
+        return "admin/editSH";
+    }
 
-    private void setModelAttributes(Model model, Date date, List<Employee> em, List<Shift> sh) {
-        model.addAttribute("searchEm", new Search());
+
+    private void setModelAttributes(Model model, Search search, Date date, List<Employee> em, List<Shift> sh) {
+        model.addAttribute("searchEm", search);
         model.addAttribute("date", date);
         model.addAttribute("employees", em);
         model.addAttribute("shifts", sh);
     }
+
+    private void setModelAttributes(Model model, List<Employee> em, List<Shift> sh) {
+        model.addAttribute("currentShift", List.of(shiftService.findByDate(LocalDate.now())));
+        model.addAttribute("searchEm", new Search());
+        model.addAttribute("date", new Date());
+        model.addAttribute("employees", em);
+        model.addAttribute("shifts", sh);
+    }
+
+//    private void setModelForShiftEdit(Model model, Shift shift) {
+//        model.addAttribute("employees", employeeService.findByName().stream()
+//                .map(em -> em[0] + " " + em[1] + " " + em[2] + " " + em[3]).toList());
+//        model.addAttribute("shift", shift);
+//    }
 
     @Getter
     @Setter
